+++ src/main/kotlin/glsl/plugin/psi/named/types/builtins/GlslBuiltinType.kt
@@ -1,0 +1,3 @@
+package glsl.plugin.psi.named.types.builtins
+
+interface GlslBuiltinType