+++ src/test/kotlin/GlslDummyTest.kt
@@ -9,10 +9,10 @@ class GlslDummyTest : BasePlatformTestCase() {
     }
 
     fun testDummy() {
-//        myFixture.configureByFiles("dummy.glsl")
-//        myFixture.checkHighlighting(false, false, false)
-        val reference = myFixture.getReferenceAtCaretPosition("dummy.glsl")
-        val resolve = reference?.resolve()
-        assertNotNull(resolve)
+        myFixture.configureByFiles("dummy.glsl")
+        myFixture.checkHighlighting(false, false, false)
+//        val reference = myFixture.getReferenceAtCaretPosition("dummy.glsl")
+//        val resolve = reference?.resolve()
+//        assertNotNull(resolve)
     }
 }