+++ src/test/testData/annotator/AnnotatorFile4.glsl
@@ -1,2 +1,2 @@
-#define <error descr="MACRO_FUNCTION or MACRO_OBJECT expected, got '<'"><</error>info A</info> 0.70
+#define <error descr="MACRO_FUNCTION or MACRO_OBJECT expected, got '<'"><</error>info A 0.70
 int a9 = vec3(1.0, A, 1.0);